Move Metadata types into their own module
